# hms-api-documents.yaml
openapi: 3.0.3
info:
  title: Hospital Management System - OpenAPI 3.0
  description: |-
    Tài liệu API của Hệ thống quản lý bệnh viện (Hospital Management System)
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /staff/doctors/{doctor_id}:
    get:
      tags:
        - doctors
      summary: Xem thông tin bác sĩ
      description: Xem thông tin của 1 bác sĩ theo id
      operationId: getDoctorById
      parameters:
        - name: doctor_id
          in: path
          description: "Id của bác sĩ cần xem chi tiết. Dùng cgRXxtrmo545vI5YbAZH để test"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OneDoctor"
            application/xml:
              schema:
                $ref: "#/components/schemas/OneDoctor"
        "400":
          description: Invalid doctor_id supplied
        "404":
          description: Doctor not found
    put:
      tags:
        - doctors
      summary: Cập nhật thông tin của bác sĩ
      description: Cập nhật thông tin của bác sĩ
      operationId: updateDoctorById
      parameters:
        - name: doctor_id
          in: path
          description: "Id của bác sĩ cần cập nhật thông tin. Dùng cgRXxtrmo545vI5YbAZH để test"
          required: true
          schema:
            type: string
      requestBody:
        description: Cập nhật thông tin của bác sĩ
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDoctor"
          application/xml:
            schema:
              $ref: "#/components/schemas/AddDoctor"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AddDoctor"
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid doctor_id supplied
        "404":
          description: Doctor not found
    delete:
      tags:
        - doctors
      summary: Xoá bác sĩ
      description: Xoá bác sĩ khỏi cơ sở dữ liệu
      operationId: deleteDoctorById
      parameters:
        - name: doctor_id
          in: path
          description: Id của bác sĩ cần xoá. **Tạo 1 bác sĩ mới rồi lên firebase lấy id để xoá.**
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid doctor_id
  /staff/doctors/{doctor_id}/patients:
    get:
      tags:
        - doctors
      summary: Xem bệnh nhân của bác sĩ
      description: Xem tất cả bệnh nhân của bác sĩ
      operationId: getDoctorPatients
      parameters:
        - name: doctor_id
          in: path
          description: "Id của bác sĩ cần xem danh sách bệnh nhân. Dùng cgRXxtrmo545vI5YbAZH để test"
          required: true
          schema:
            type: string
        - name: pageSize
          in: query
          description: Số lượng phần tử trong 1 trang. *Mặc định là 10*
          required: false
          schema:
            type: string
        - name: currentPage
          in: query
          description: Trang hiện tại. *Mặc định là 1*
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Invalid doctor_id
  /staff/doctors/departments/statistics:
    get:
      tags:
        - doctors
      summary: Thống kê bác sĩ theo từng khoa
      description: Thống kê số lượng bác sĩ theo từng khoa
      operationId: countDoctorByDepartment
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountByDepartment"
        "404":
          description: Doctor not found
...
  /patients/{patient_id}:
    get:
      tags:
        - patients
      summary: Xem thông tin bệnh nhân
      description: Xem thông tin của 1 bệnh nhân
      operationId: getPatientById
      parameters:
        - name: patient_id
          in: path
          description: "Id của bệnh nhân cần xem chi tiết. Dùng 0n1HhIX3QI8GmJi8qfwK để test"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnePatient"
            application/xml:
              schema:
                $ref: "#/components/schemas/OnePatient"
        "400":
          description: Invalid patient_id supplied
        "404":
          description: Patient not found
    put:
      tags:
        - patients
      summary: Cập nhật thông tin của bệnh nhân
      description: Cập nhật thông tin của bệnh nhân
      operationId: updatePatientById
      parameters:
        - name: patient_id
          in: path
          description: "Id của bệnh nhân cần cập nhật thông tin. Dùng 0n1HhIX3QI8GmJi8qfwK để test"
          required: true
          schema:
            type: string
      requestBody:
        description: Cập nhật thông tin của bệnh nhân
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPatient"
          application/xml:
            schema:
              $ref: "#/components/schemas/AddPatient"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AddPatient"
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid patient_id supplied
        "404":
          description: Patient not found
    delete:
      tags:
        - patients
      summary: Xoá bệnh nhân
      description: Xoá bệnh nhân khỏi cơ sở dữ liệu
      operationId: deletePatientById
      parameters:
        - name: patient_id
          in: path
          description: Id của bệnh nhân cần xoá. **Tạo 1 bệnh nhân mới rồi lên firebase lấy id để xoá.**
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid patient_id
...
  /patients/departments/statistics:
    get:
      tags:
        - patients
      summary: Thống kê bệnh nhân theo từng khoa
      description: Thống kê số lượng bệnh nhân theo từng khoa
      operationId: countPatientByDepartment
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountByDepartment"
        "404":
          description: Patient not found
...
  /staff/nurse/{nurse_id}:
    get:
      tags:
        - nurses
      summary: Xem thông tin y tá
      description: Xem thông tin y tá theo id
      operationId: getNurseById
      parameters:
        - name: nurse_id
          in: path
          description: "Id của y tá cần xem chi tiết. Dùng Y6Z3vWVhIJhJSiM7od4R để test"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OneNurse"
            application/xml:
              schema:
                $ref: "#/components/schemas/OneNurse"
        "400":
          description: Invalid nurse_id supplied
        "404":
          description: Nurse not found
    put:
      tags:
        - nurses
      summary: Cập nhật thông tin của y tá
      description: Cập nhật thông tin của y tá
      operationId: updateNurseById
      parameters:
        - name: nurse_id
          in: path
          description: "Id của y tá cần cập nhật. Dùng Y6Z3vWVhIJhJSiM7od4R để test"
          required: true
          schema:
            type: string
      requestBody:
        description: Cập nhật thông tin của y tá
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddNurse"
          application/xml:
            schema:
              $ref: "#/components/schemas/AddNurse"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AddNurse"
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid nurse_id supplied
        "404":
          description: Nurse not found
    delete:
      tags:
        - nurses
      summary: Xóa y tá
      description: Xóa y tá khỏi cơ sở dữ liệu
      operationId: deleteNurseById
      parameters:
        - name: nurse_id
          in: path
          description: Id của y tá cần xoá. **Tạo 1 y tá mới rồi lên firebase lấy id để xoá.**
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid nurse_id
...
components:
  schemas:
...
    OneDoctor:
      type: object
      properties:
        id:
          type: string
          example: cgRXxtrmo545vI5YbAZH
        firstName:
          type: string
          example: Phuc
        lastName:
          type: string
          example: Le
        dateOfBirth:
          type: string
          format: date
          example: 05/03/2002
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "0123456789"
        salary:
          type: integer
          example: 15000000
        department:
          type: string
          example: Khoa nhi
        absence:
          type: boolean
          example: false
    AddDoctor:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - salary
        - department
        - absence
      type: object
      properties:
        firstName:
          type: string
          example: K
        lastName:
          type: string
          example: Nguyen
        dateOfBirth:
          type: string
          format: date
          example: 12/31/1994
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "0004567891"
        salary:
          type: integer
          example: 20000000
        department:
          type: string
          example: Khoa nội
        absence:
          type: boolean
          example: true
    Patient:
      type: object
      properties:
        patients:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 0n1HhIX3QI8GmJi8qfwK
              firstName:
                type: string
                example: Thinh
              lastName:
                type: string
                example: Nguyen Thi
              gender:
                type: string
                example: male
              phoneNumber:
                type: string
                example: "6410983572"
              healthInsurance:
                type: string
                example: "true"
              doctorResponbility:
                type: string
                example: "cgRXxtrmo545vI5YbAZH"
              dateOfBirth:
                type: string
                format: date
                example: 18/04/2008
        pageSize:
          type: integer
          example: 10
        currentPage:
          type: integer
          example: 1
        totalPage:
          type: integer
          example: 1
    OnePatient:
      type: object
      properties:
        id:
          type: string
          example: 0n1HhIX3QI8GmJi8qfwK
        firstName:
          type: string
          example: Thinh
        lastName:
          type: string
          example: Nguyen Thi
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "6410983572"
        healthInsurance:
          type: string
          example: "true"
        doctorResponbility:
          type: string
          example: "cgRXxtrmo545vI5YbAZH"
        dateOfBirth:
          type: string
          format: date
          example: 18/04/2008
    AddPatient:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - healthInsurance
        - doctorResponbility
      type: object
      properties:
        firstName:
          type: string
          example: A
        lastName:
          type: string
          example: Nguyen
        dateOfBirth:
          type: string
          format: date
          example: 17/5/2002
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "0949777565"
        healthInsurance:
          type: boolean
          example: true
        doctorResponsibility:
          type: string
          example: HNC4vvn6DLES09emJQBl
...
    OneNurse:
      type: object
      properties:
        id:
          type: string
          example: Y6Z3vWVhIJhJSiM7od4R
        firstName:
          type: string
          example: Phuc
        lastName:
          type: string
          example: Vien
        dateOfBirth:
          type: string
          format: date
          example: 12/01/2002
        gender:
          type: string
          example: female
        phoneNumber:
          type: string
          example: "9870123456"
        salary:
          type: integer
          example: 30000000
        absence:
          type: boolean
          example: false
    AddNurse:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - salary
        - absence
      type: object
      properties:
        firstName:
          type: string
          example: S
        lastName:
          type: string
          example: Nghiem
        dateOfBirth:
          type: string
          format: date
          example: 12/01/1998
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "00110007238"
        salary:
          type: integer
          example: 20000000
        absence:
          type: boolean
          example: true
...
    CountByDepartment:
      type: object
      properties:
        departments:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              total:
                type: integer
          example:
            - name: Khoa nội
              total: 1
            - name: Khoa ngoại
              total: 1
            - name: Khoa nhi
              total: 1
            - name: Khoa sản
              total: 1
            - name: Khoa mắt
              total: 1